+++ e2e/flows/dependencies-as-packages.e2e.2.ts
@@ -311,6 +311,7 @@ chai.use(require('chai-fs'));
       });
       describe('installing in Harmony', () => {
         before(() => {
+          helper.command.resetFeatures();
           helper.scopeHelper.reInitLocalScopeHarmony();
           helper.scopeHelper.addRemoteScope();
           helper.command.runCmd('npm init -y');
