+++ e2e/flows/out-of-sync-componets.e2e.3.ts
@@ -12,7 +12,6 @@ describe('components that are not synced between the scope and the consumer', fu
   let helper: Helper;
   before(() => {
     helper = new Helper();
-    helper.command.setFeatures('legacy-workspace-config');
   });
   after(() => {
     helper.scopeHelper.destroy();
@@ -100,40 +99,6 @@ describe('components that are not synced between the scope and the consumer', fu
         const newId = `${helper.scopes.remote}/bar/foo@0.0.1`;
         expect(bitMap).to.have.property(newId);
         expect(bitMap[newId].exported).to.be.true;
-      });
-    });
-  });
-  describe('consumer with a new component and scope with the same component as exported with defaultScope configured', () => {
-    let scopeOutOfSync;
-    before(() => {
-      helper.scopeHelper.setNewLocalAndRemoteScopesHarmony();
-      helper.bitJsonc.setupDefault();
-      helper.fixtures.createComponentBarFoo();
-      helper.fixtures.addComponentBarFooAsDir();
-      const bitMap = helper.bitMap.read();
-      helper.command.tagAllWithoutBuild();
-      helper.command.export();
-      // the mimic and import here is to make sure the local doesn't have the symlink object
-      helper.git.mimicGitCloneLocalProjectHarmony();
-      helper.scopeHelper.addRemoteScope();
-      helper.command.importAllComponents();
-      helper.bitMap.write(bitMap);
-      scopeOutOfSync = helper.scopeHelper.cloneLocalScope();
-    });
-    describe('bit tag', () => {
-      it('should tag the component to the next version of what the scope has', () => {
-        const output = helper.command.runCmd('bit tag bar/foo --force --patch');
-        expect(output).to.have.string('0.0.2');
-      });
-    });
-    describe('bit status', () => {
-      before(() => {
-        helper.scopeHelper.getClonedLocalScope(scopeOutOfSync);
-        helper.command.status();
-      });
-      it('should sync .bitmap according to the scope', () => {
-        helper.command.expectStatusToBeClean();
-        helper.bitMap.expectToHaveIdHarmony('bar/foo', '0.0.1', helper.scopes.remote);
       });
     });
   });
