+++ e2e/harmony/out-of-sync-components-harmony.e2e.ts
@@ -1,0 +1,50 @@
+import chai, { expect } from 'chai';
+import Helper from '../../src/e2e-helper/e2e-helper';
+
+chai.use(require('chai-fs'));
+
+describe('components that are not synced between the scope and the consumer', function () {
+  this.timeout(0);
+  let helper: Helper;
+  before(() => {
+    helper = new Helper();
+    helper.command.setFeatures('legacy-workspace-config');
+  });
+  after(() => {
+    helper.scopeHelper.destroy();
+  });
+  describe('consumer with a new component and scope with the same component as exported with defaultScope configured', () => {
+    let scopeOutOfSync;
+    before(() => {
+      helper.scopeHelper.setNewLocalAndRemoteScopesHarmony();
+      helper.bitJsonc.setupDefault();
+      helper.fixtures.createComponentBarFoo();
+      helper.fixtures.addComponentBarFooAsDir();
+      const bitMap = helper.bitMap.read();
+      helper.command.tagAllWithoutBuild();
+      helper.command.export();
+      // the mimic and import here is to make sure the local doesn't have the symlink object
+      helper.git.mimicGitCloneLocalProjectHarmony();
+      helper.scopeHelper.addRemoteScope();
+      helper.command.importAllComponents();
+      helper.bitMap.write(bitMap);
+      scopeOutOfSync = helper.scopeHelper.cloneLocalScope();
+    });
+    describe('bit tag', () => {
+      it('should tag the component to the next version of what the scope has', () => {
+        const output = helper.command.runCmd('bit tag bar/foo --force --patch');
+        expect(output).to.have.string('0.0.2');
+      });
+    });
+    describe('bit status', () => {
+      before(() => {
+        helper.scopeHelper.getClonedLocalScope(scopeOutOfSync);
+        helper.command.status();
+      });
+      it('should sync .bitmap according to the scope', () => {
+        helper.command.expectStatusToBeClean();
+        helper.bitMap.expectToHaveIdHarmony('bar/foo', '0.0.1', helper.scopes.remote);
+      });
+    });
+  });
+});
